// src/types.ts
export interface Draw {
  id: string;
  date: string;
  numbers: number[]; // 3 numbers, each 1-6
  isManual?: boolean;
  raw?: any;
}

export type SumType = "TAI" | "XIU" | "HOA";

export interface ProbabilityScores {
  TAI: number;
  XIU: number;
  HOA: number;
}